==> 3_Kidos_Toestemmingen_Schoolgaanden.ipynb <==
### 1. Import modules

import warnings
warnings.filterwarnings('ignore')

import pyodbc
import pandas as pd
import numpy as np

### 2. Import data to Pandas dataframe (Kidos_Toestemmingen)

# Import pickles
df_schoolgaanden = pd.read_pickle('4_Data\Pickles\schoolgaanden_observaties.pkl')

# Set to True or False depending on whether you want to import new data
NEWIMPORT = False

if NEWIMPORT:
    # Replace server name if source changes
    server = 'AZ-ODB0\ODBWB'
    database = 'OnderzoeksDB_WB'

    # Create a connection string
    connection_string = f'DRIVER={{SQL Server}};SERVER={server};DATABASE={database}'

    # Establish connection
    connection = pyodbc.connect(connection_string)

    # Define columns to select
    selected_columns = ['FK_CLIENT_ID',
                        'TOESTEMMINGSVRAAG',
                        'ANTWOORD',
                        'BEGINDATUM']

    # Define a list of Clientnummers to include
    included_clientnummers = df_schoolgaanden['Clientnummer'].unique()
    print(len(included_clientnummers))

    # Create SQL query string with specific columns and inclusion criteria
    query = f'SELECT {', '.join(selected_columns)} FROM Kidos_TOESTEMMINGEN WHERE FK_CLIENT_ID IN ({', '.join(included_clientnummers)})'

    # Execute the query and fetch the data into a Pandas DataFrame
    df_toestemmingen_og = pd.read_sql(query, connection)

    # Close the connection
    connection.close()

    # Save df_contactmomenten_og as a pickle
    df_toestemmingen_og.to_pickle('4_Data\Pickles\\toestemmingen_schoolgaanden_og.pkl')
else:
    # Load the pickle
    df_toestemmingen_og = pd.read_pickle('4_Data\Pickles\\toestemmingen_schoolgaanden_og.pkl')

### 3. Process

# Store copy of unique records of df_og in df
df_toestemmingen = df_toestemmingen_og.drop_duplicates()

print(f'The amount of relations in the df_contactmomenten is: {len(df_toestemmingen)}, the amount of unique client ID\'s is: {len(df_toestemmingen['FK_CLIENT_ID'].unique())}')

# Remove all rows where Observatie_waarde != 'Toestemming gegevensuitwisseling RVP'
toestemmingen_rvp = df_toestemmingen[df_toestemmingen['TOESTEMMINGSVRAAG'] == 'Toestemming gegevensuitwisseling RVP']

# Get most recent 'toestemmingen'
schoolgaanden_toestemmingen_rvp = toestemmingen_rvp.groupby('FK_CLIENT_ID')['BEGINDATUM'].idxmax()

# Get all the indices (.values) from the df
most_recent = toestemmingen_rvp.loc[schoolgaanden_toestemmingen_rvp.values]

# Select only constant variables, observaties will be added later
schoolgaanden_toestemmingen = most_recent[['FK_CLIENT_ID', 'ANTWOORD']]

# In column ANTWOORD replace 'wel toestemming' with 1 and 'geen toestemming' with 0
schoolgaanden_toestemmingen['ANTWOORD'] = np.where(schoolgaanden_toestemmingen['ANTWOORD'] == 'wel toestemming', 1, 0)

# Get the ID's and ANTWOORD and rename the columns
schoolgaanden_toestemmingen = schoolgaanden_toestemmingen[['FK_CLIENT_ID', 'ANTWOORD']]
schoolgaanden_toestemmingen.columns = ['Clientnummer', 'Toestemming']

print(f'The amount of relations in the df_contactmomenten is: {len(schoolgaanden_toestemmingen)}, the amount of unique client ID\'s is: {len(schoolgaanden_toestemmingen['Clientnummer'].unique())}')

### 4. Save as .pkl

# Save schoolgaanden_toestemmingen as a pickle
schoolgaanden_toestemmingen.to_pickle('4_Data\Pickles\schoolgaanden_toestemmingen.pkl')
