==> 4_CBS_Schoolgaanden_Postcodes.ipynb <==
### 1. Import modules

import warnings
warnings.filterwarnings('ignore')

import pyodbc
import pandas as pd
import numpy as np
import xlwings as xw
import geopandas as gpd

### 2. Import schoolgaanden to Pandas dataframe

# Import pickles
df_schoolgaanden = pd.read_pickle('4_Data\Pickles\schoolgaanden_observaties.pkl')

# Save only columns PersonNr and Postcode
df_schoolgaanden_pc6 = df_schoolgaanden[['Clientnummer', 'Postcode']]

### 2. Import CBS .gpkg

#### 2.1 PC6

# # Reading GeoPackage file
# input_file = '4_Data\cbs_pc6_2022_v1.gpkg'
# cbs_df = gpd.read_file(input_file)

# # Replace every -99997 with NaN
# cbs_df = cbs_df.replace(-99997, np.nan)

# # Rename postcode to Postcode
# cbs_df.rename(columns={'postcode':'Postcode'}, inplace=True)

# # Merge schoolgaanden_pc6 with cbs_df on postcode, preserving all columns in schoolgaanden_pc6
# schoolgaanden_cbs6 = pd.merge(df_schoolgaanden_pc6, cbs_df, on='Postcode', how='left')

# # Drop columns postcode and geometry
# schoolgaanden_cbs6 = schoolgaanden_cbs6.drop(columns=['Postcode', 'geometry'])

# # Set Clietnummer dtype to object
# schoolgaanden_cbs6['Clientnummer'] = schoolgaanden_cbs6['Clientnummer'].astype('str')

# # Print the amount of NaN values per column as percentages
# print(schoolgaanden_cbs6.isna().sum() / len(schoolgaanden_cbs6) * 100)

#### 2.2 PC4

# Reading GeoPackage file
input_file = '4_Data\cbs_pc4_2022_v1.gpkg'
cbs4_df = gpd.read_file(input_file)

# Replace every -99997 with NaN
cbs4_df = cbs4_df.replace(-99997, np.nan)

# Remove last two characters of schoolgaanden_pc6['postcode']
df_schoolgaanden['postcode4'] = df_schoolgaanden['Postcode'].str[:-2]

# Rename schoolgaanden_pc6 to schoolgaanden_pc4
schoolgaanden_pc4 = df_schoolgaanden[['Clientnummer', 'postcode4']]

# Merge schoolgaanden_pc6 with cbs_df on postcode, preserving all columns in schoolgaanden_pc6
schoolgaanden_cbs4 = pd.merge(schoolgaanden_pc4, cbs4_df, on='postcode4', how='left')

# Drop columns postcode and geometry
schoolgaanden_cbs4 = schoolgaanden_cbs4.drop(columns=['postcode4', 'geometry'])

# Set Clietnummer dtype to object
schoolgaanden_cbs4['Clientnummer'] = schoolgaanden_cbs4['Clientnummer'].astype('str')

# Print the amount of NaN values per column as percentages
print(schoolgaanden_cbs4.isna().sum() / len(schoolgaanden_cbs4) * 100)

### 5. Save .pkl's

# Save as a pickle
schoolgaanden_cbs4.to_pickle('4_Data\Pickles\schoolgaanden_cbs4.pkl')
# schoolgaanden_cbs6.to_pickle('4_Data\Pickles\schoolgaanden_cbs6.pkl')
