==> 4_CBS_Zuigelingen_Postcodes.ipynb <==
### 1. Import modules

import warnings
warnings.filterwarnings('ignore')

import pyodbc
import pandas as pd
import numpy as np
import geopandas as gpd

# Import pickles
zuigelingen_pc6 = pd.read_pickle('4_Data\Pickles\zuigelingen_observaties.pkl')

### 2. Import CBS .gpkg

#### 2.1 PC6

# # Reading GeoPackage file
# input_file = '4_Data\cbs_pc6_2022_v1.gpkg'
# cbs_df = gpd.read_file(input_file)

# # Replace every -99997 with NaN
# cbs_df = cbs_df.replace(-99997, np.nan)

# example = ['4884JA', '4884JD', '5111HH', '4264AL']
# cbs_df[cbs_df['postcode'].isin(example)][['postcode', 'percentage_geb_buiten_nederland_herkomst_europa']]

# # Rename postcode to Postcode6_BRP
# cbs_df.rename(columns={'postcode':'PC6'}, inplace=True)

# # Merge zuigelingen_pc6 with cbs_df on postcode, preserving all columns in zuigelingen_pc6
# zuigelingen_cbs6 = pd.merge(zuigelingen_pc6[['PC6', 'Clientnummer']], cbs_df, on='PC6', how='left')

# # Drop columns postcode and geometry
# zuigelingen_cbs6 = zuigelingen_cbs6.drop(columns=['PC6', 'geometry'])

# # Set Clietnummer dtype to object
# zuigelingen_cbs6['Clientnummer'] = zuigelingen_cbs6['Clientnummer'].astype('str')

# # Print the amount of NaN values per column as percentages
# print(zuigelingen_cbs6.isna().sum() / len(zuigelingen_cbs6) * 100)

#### 2.2 PC4

# Reading GeoPackage file
input_file = '4_Data\cbs_pc4_2022_v1.gpkg'
cbs4_df = gpd.read_file(input_file)

# Replace every -99997 with NaN
cbs4_df = cbs4_df.replace(-99997, np.nan)

# Rename postcode to Postcode4_BRP
cbs4_df.rename(columns={'postcode4':'PC4'}, inplace=True)

# Rename zuigelingen_pc6 to zuigelingen_pc4
zuigelingen_pc4 = zuigelingen_pc6[['Clientnummer', 'PC4']]

# Merge zuigelingen_pc6 with cbs_df on postcode, preserving all columns in zuigelingen_pc6
zuigelingen_cbs4 = pd.merge(zuigelingen_pc4, cbs4_df, on='PC4', how='left')

# Drop columns postcode and geometry
zuigelingen_cbs4 = zuigelingen_cbs4.drop(columns=['PC4', 'geometry'])

# Set Clietnummer dtype to object
zuigelingen_cbs4['Clientnummer'] = zuigelingen_cbs4['Clientnummer'].astype('str')

# Print the amount of NaN values per column as percentages
print(zuigelingen_cbs4.isna().sum() / len(zuigelingen_cbs4) * 100)

### 5. Save .pkl's

# Save as a pickle
zuigelingen_cbs4.to_pickle('4_Data\Pickles\zuigelingen_cbs4.pkl')
# zuigelingen_cbs6.to_pickle('4_Data\Pickles\zuigelingen_cbs6.pkl')
