==> 5_CBS_Schoolgaanden_Buurten.ipynb <==
### 1. Import modules

import warnings
warnings.filterwarnings('ignore')

import pandas as pd

### 2. Import schoolgaanden to Pandas dataframe

# Import pickles
df_schoolgaanden = pd.read_pickle('4_Data\Pickles\schoolgaanden_observaties.pkl')

# Save only columns PersonNr and Postcode
df_schoolgaanden_buurt = df_schoolgaanden[['Clientnummer', 'CBS_Buurtcode_BRP']]

# Print missingness percentages in the data
print(df_schoolgaanden_buurt.isnull().mean())

### 3. Import CBS .csv with socio-demographic variables

# Reading csv to pandas dataframe
input_file = '4_Data\Kerncijfers_wijken_en_buurten_2022_25062024_152855.csv'
cbs_buurten = pd.read_csv(input_file, delimiter=';')

cbs_buurten

# Rename Regioaanduiding/Codering (code) to CBS_Buurtcode_BRP
cbs_buurten.rename(columns={'Regioaanduiding/Codering (code)': 'CBS_Buurtcode_BRP'}, inplace=True)

# From CBS_Buurtcode_BRP remove first two characters if they are 'BU'
cbs_buurten['CBS_Buurtcode_BRP'] = cbs_buurten['CBS_Buurtcode_BRP'].apply(lambda x: x[2:] if x[:2] == 'BU' else x)

# Merge
schoolgaanden_buurt = pd.merge(df_schoolgaanden_buurt, cbs_buurten, on='CBS_Buurtcode_BRP', how='left')

# Drop columns postcode and geometry
schoolgaanden_buurt = schoolgaanden_buurt.drop(columns=['CBS_Buurtcode_BRP', 'Wijken en buurten'])

# Set Clietnummer dtype to object
schoolgaanden_buurt['Clientnummer'] = schoolgaanden_buurt['Clientnummer'].astype('str')

# Print the amount of NaN values per column as percentages
print(schoolgaanden_buurt.isna().mean())

### 4. Save .pkl's

# Save as a pickle
schoolgaanden_buurt.to_pickle('4_Data\Pickles\schoolgaanden_buurt.pkl')
