==> 5_CBS_Zuigelingen_Buurten.ipynb <==
### 1. Import modules

import warnings
warnings.filterwarnings('ignore')

import pandas as pd

### 2. Import schoolgaanden to Pandas dataframe

# Import pickles
df_zuigelingen = pd.read_pickle('4_Data\Pickles\zuigelingen_observaties.pkl')

# Save only columns PersonNr and Postcode
df_zuigelingen_buurt = df_zuigelingen[['Clientnummer', 'BuurtCode']]

# Print missingness percentages in the data
print(df_zuigelingen_buurt.isnull().mean())

### 3. Import CBS .csv

# Reading csv to pandas dataframe
input_file = '4_Data\Kerncijfers_wijken_en_buurten_2022_25062024_152855.csv'
cbs_df = pd.read_csv(input_file, delimiter=';')

# Rename Regioaanduiding/Codering (code) to CBS_Buurtcode_BRP
cbs_df.rename(columns={'Regioaanduiding/Codering (code)': 'BuurtCode'}, inplace=True)

# From CBS_Buurtcode_BRP remove first two characters if they are 'BU'
cbs_df['BuurtCode'] = cbs_df['BuurtCode'].apply(lambda x: x[2:] if x[:2] == 'BU' else x)

# Merge
zuigelingen_buurt = pd.merge(df_zuigelingen_buurt, cbs_df, on='BuurtCode', how='left')

# Drop columns postcode and geometry
zuigelingen_buurt = zuigelingen_buurt.drop(columns=['BuurtCode', 'Wijken en buurten'])

# Set Clietnummer dtype to object
zuigelingen_buurt['Clientnummer'] = zuigelingen_buurt['Clientnummer'].astype('str')

# Print the amount of NaN values per column as percentages
print(zuigelingen_buurt.isna().sum() / len(zuigelingen_buurt) * 100)

len(zuigelingen_buurt)

### 4. Save .pkl's

# Save as a pickle
zuigelingen_buurt.to_pickle('4_Data\Pickles\zuigelingen_buurt.pkl')
