==> 6_CoronIT_Vaccinaties_Schoolgaanden.ipynb <==
### 1. Import modules

import warnings
warnings.filterwarnings('ignore')

import pyodbc
import pandas as pd
import numpy as np

### 2. Import data to Pandas dataframe (CoronIT)

# Import pickles
df_coronit = pd.read_csv('4_Data\CoronIT\ouders_schoolgaand_vaccinatie.csv')

### 3. Process

#### 3.1 Clean

# Change all 'MAN' to 'Mannelijk' and 'VROUW' to 'Vrouwelijk' in column geslacht
df_coronit['geslacht'] = df_coronit['geslacht'].replace({'MAN': 'mannelijk', 'VROUW': 'vrouwelijk'})

#### 3.2 Handle duplicates

len(df_coronit)

# Drop all rows that are exact duplicates
df_coronit = df_coronit.drop_duplicates()

len(df_coronit)

# Drop all rows that are duplicates but have one row with empty geslacht
df_coronit = df_coronit[~((df_coronit.duplicated(subset=['postcode', 'huisnummer', 'geboortedatum_ym', 'gevaccineerd'], keep=False)) & (df_coronit['geslacht'].isnull()))]

len(df_coronit)

# Drop all rows that are duplicates but have one row with empty geslacht
df_coronit = df_coronit[~((df_coronit.duplicated(subset=['postcode', 'huisnummer', 'geboortedatum_ym'], keep=False)) & (df_coronit['geslacht'].isnull()))]

len(df_coronit)

# Check if there are duplicates left
len(df_coronit[df_coronit.duplicated(subset=['postcode', 'huisnummer', 'geboortedatum_ym', 'geslacht'], keep=False) == True])

### 4. Save as .pkl

# Save zuigelingen_toestemmingen as a pickle
df_coronit.to_pickle('4_Data\Pickles\schoolgaanden_coronit.pkl')
