==> 9_UMAP15_Schoolgaanden_Notvaccinated.ipynb <==
### 1. Import modules

import warnings
warnings.filterwarnings('ignore')

import pandas as pd
import numpy as np
import pickle
import matplotlib.pyplot as plt
import umap
import umap.plot

### 2. Merge .pkl's

# Load data
X_schoolgaanden_numericals = pd.read_pickle('../4_Data/Pickles/X_schoolgaanden_notvaccinated_numericals.pkl')
X_schoolgaanden_booleans = pd.read_pickle('../4_Data/Pickles/X_schoolgaanden_notvaccinated_booleans.pkl')

X_schoolgaanden_numericals.dtypes

X_schoolgaanden_booleans.dtypes

### 3. UMAP - Visualization of numerical space

#### 3.2 Model

# Save mappers
numeric_mapper1 = umap.UMAP(metric='euclidean', n_neighbors=100, min_dist=0.1, n_components=15, random_state=42).fit(X_schoolgaanden_numericals)
numeric_mapper2 = umap.UMAP(metric='euclidean', n_neighbors=100, min_dist=0.1, n_components=15, random_state=43).fit(X_schoolgaanden_numericals)

### 4. UMAP - Visualization of boolean space

#### 4.2 Model

# Save mappers
boolean_mapper1 = umap.UMAP(metric='dice', n_neighbors=100, min_dist=0.1, n_components=15, random_state=42).fit(X_schoolgaanden_booleans)
boolean_mapper2 = umap.UMAP(metric='dice', n_neighbors=100, min_dist=0.1, n_components=15, random_state=43).fit(X_schoolgaanden_booleans)

### 5. Combine UMAP models

# Plot
intersection_mapper1 = numeric_mapper1 * boolean_mapper1
intersection_mapper2 = numeric_mapper2 * boolean_mapper2

intersection_embedding1 = intersection_mapper1.embedding_
intersection_embedding2 = intersection_mapper2.embedding_

# Plot
union_mapper1 = numeric_mapper1 + boolean_mapper1
union_mapper2 = numeric_mapper2 + boolean_mapper2

union_embedding1 = union_mapper1.embedding_
union_embedding2 = union_mapper2.embedding_

### 6. Save as arrays

# Save embeddings
np.save('../4_Data/Arrays/X_schoolgaanden_notvaccinated_UMAP15_intersection1.npy', intersection_embedding1)
np.save('../4_Data/Arrays/X_schoolgaanden_notvaccinated_UMAP15_intersection2.npy', intersection_embedding2)
np.save('../4_Data/Arrays/X_schoolgaanden_notvaccinated_UMAP15_union1.npy', union_embedding1)
np.save('../4_Data/Arrays/X_schoolgaanden_notvaccinated_UMAP15_union2.npy', union_embedding2)
